+++ Populations/Rules/BorderTotalistic.cs
@@ -30,7 +30,7 @@ namespace CellularAutomata.Populations
 		}  // end Parse, public void method
 
 		// implements rule, returns new state value
-		public int Implement (ICell cell)
+		public ushort Implement (ICell cell)
 		{
 
 			long state;
@@ -48,7 +48,7 @@ namespace CellularAutomata.Populations
 			placeValue = placeValue / color;
 			state = state / placeValue;
 
-			return (int)state;
+			return (ushort)state;
 
 		} // end Implement, public virtual int method
